Introduced PWSAuxCore
PWSAuxCore is a subclass of PWScore who's sole purpose is to skip
loading the database's preference string into the PWSPref singleton.
This saves the save/restore cycle when using a second core to read
a file.
The alternative would be to make the pref a core member instead of
a singleton. It's debatable which approach is preferable, but
this is certainly a smaller code change.