[1067] Fail gracefully when trying to encrypt a file > 2GB.
Note that supporting file encryption for files > 2GB is not doable
without breaking backwards compatability, since we encrypt the length
of the data first using 32 bits. Fixable, but I wondier if it's worth
the effort. In any case, I've changed the file length type to be 64 bits,
(a) to detect this gracefully, and (b) towards supporting password
databases that may exceed 2GB, e.g., if/when we have attachments.
(Note that the 2G limit is inherent on a single field, not the file...)