Dragbar no longer copies to clipboard at all
rather than copying and clearing, which is effectively useless.
This is because a user correctly pointed out that using the clipboard
in addition is an inherent securty risk, and I was unaware that the
basic functionality is provided w/o the clipboard at all.

(Noted by Andrey during code review)