[BR1201] Fix Yubikey passphrase inompatability
Turns out that Linux and Windows were using the passphrase differently
to generate the Yubikey challenge. This caused a Yubikey-protected
database created on one platform to be unopenable on the other.

Gory details are in CYubiMixin::PerformChallengeResponse().

From the user perspective: From this commit on, challenges are generated
"windows style" unless SHIFT is press while clicking on the Yubikey dialog button.
This allows maximum flexibilty: opening old and new, converting from one to the other
and creating a database in either method.

Note that the default behavior is for Windows compatablity, NOT backwards
compatability.