V4 support for multiple passwords
Passes unit test as well as maintains bwd compat
with single user use-case, i.e., no need to modify callers (PWScore).

Need to seriously consider adding UUID to key block, so as to allow
association with arbitrary data that can be encoded in header, such
as the user name associated with a given keyblock.