[FR-789] Tweak S/DCA edit/view sematics for shortcuts
The DCA/Shift+DCA normal process is to use the entry's action unless it
has not been set, in which case the default action (application preference)
is performed. If the entry is a shortcut, then for all actions other than
View/Edit, the base entry's setting is performed. If the action was
View/Edit, then this applied to the shortcut itself. This FR changes
View/Edit action to View/Edit the base entry to be consistent with all
other actions. A shortcut can now only be viewed/edited via the Edit menu
or the right click context menu.

Signed-off-by: ronys <ronys@pwsafe.org>