Protect against use-after-free of CItem
Consider a CItem-derived instance referred after delete -
the Blowfish member will be used for Getting a field, but it will return
garbage since the internal state's been trashed. This patch tries to
the issue.

No direct evedince that this happens, but it's a possible explanation
for the rare data corruptions that we're seeing in the field since
releasing the version with lazy Blowfish creation.