Fix type mismatch in ReadCBC/WriteCBC
The CBC is updated with a buffer and its size, however, the size that
was previously being sent in was converted to a signed int, which may
not have been the same as the actual size.
The new usage is consistent with how CBC is updated in other places
in the code.