Fix compile error on 32-bit FreeBSD
There are multiple functions named Matches() that differ only by
parameter type. In 32-bit BSD, the params for two of these are
both actually int. No callers are depending on the correct
overload to be bound based on type, so the easy fix is to rename one of
them, and change the call site to use the new name.