Cleanup file lock/unlock semantics
Removed lock count which allowed same process to lock a file more than once.
I suspect this was a workaround for the config and dbase using the same lock,
which was fixed be changing the lock file name of a config file.

In any case, non-Windows didn't adhere to the counting semantics, so we're
also consistent.