Don't link pwsafe-cli with X libraries
To let it run in headless environments.

X libraries were required because core/XMLprefs linked with
a (basically no-op) function in "os" that brought in the entire
module along with all its X dependencies. The solution is to split
that no-op function into its own compilation unit so that the linker
doesn't add the module with X dependencies to the final executable.