update README.LINUX.DEVELOPERS.txt for debian
The original instructions simply point developers to a page that neglects to instruct users to obtain the relevant public key for the repository, and also there is a tendency for users to follow that page entirely and install too many packages.  Detailed instructions are now given.  (although perhaps my use of `aptitude` should be replaced with `apt-get` for the general public.. not sure)