Fix Cast-related Issues for Big-Endian Systems
It appears that a reinterpret cast of a 64-bit value down to a 32-bit
value selects the leftmost dword. On little-endian systems, this works
fine, however, on big-endian systems this results in a value of zero for
any value less than 32-bits.

Replacing the problematic reinterpret casts with static casts resolves
the issue, without affecting l-e functionality.