Fix BlowFish Compatibility on Big-Endian Systems
Encryption/decryption done on big-endian systems does not pass test
vectors and is only usable by other big-endian systems.

Based on the following note in the test code -

 * PasswordSafe Note: To pass these tests, I had to change the
 * order of the bytes in the plaintext and ciphertext vectors,
 * from {3,2,1,0,7,6,5,4} to {0,1,2,3,4,5,6,7}. This is either a
 * (cryptographically benign) bug in the implementation, or some
 * endian-ness oddity in the test data. My guess is the former.

And this mailing list message -

 https://sourceforge.net/p/passwordsafe/mailman/message/2256242/

as well as my own testing, the fix was to swap bytes in/out of the
encrypt/decrypt functions, as well as to reverse the order in the
aword struct.