Implement save and restore of tree state
Resolves issue 1407 'PasswordSafe Linux 1.01B won't retain tree state'
that was reported on Sourceforge.

Introduced a new boolean member to passwordsafe which indicates whether
a new DB was opened. Now, whenever a new DB gets opened and tree view
shown for the first time the user preference 'TreeDisplayStatusAtOpen'
gets respected.

Additionally, used the introduced boolean member to update the 'guiinfo'
instance regarding the tree display information which was outdated after
switching the DB. The tree data from a previously opened DB was used on
the newly opened DB resp. tree and caused some debug information on the
command line that some tree items couldn't be restored.