[BR1505] crash on compare - part 1
When "other" DB has a shortcut and user wants to view it:
Since the "other" db isn't displayed, there's no GUI element that can
be used to determine if the SC is the last elmenet in a group.
However, this only interests us if we edit the SC (and change the group).
Therefore, moving this to the post-edit-and-is-modified block is enough
to protect against this crash.