Gitpod: Workaround for https://github.com/gitpod-io/gitpod/issues/1565
Concluded that gitpod using `sudo apt-get` results in a permission issue (https://github.com/gitpod-io/gitpod/issues/1565). This is a workaround for this issue.

File setup-deb-dev-env
- This file has extension `.sh`, but is using a shebang `#!/bin/bash` for bash and in previous commit was calling `sh` to call this script which might break on some environments i.e environments without shebang support compiled in kernel and systems without dynamic linking (not necesary linked to gitpod, but good practice)
- Was using `apt-get` without `-y` which would cause an infinite loop in gitpod's docker environment that doesn't have set `DEBIAN_FRONTEND=noninteractive` to workaround https://github.com/gitpod-io/gitpod/issues/1171

Shellcheck (https://github.com/koalaman/shellcheck) has been recommended to ensure code quality of shell scripts

This also makes it to not use `setup-deb-dev-env.sh` let me know if you want to use this file in gitpod (seems to me as bad idea since it would be harder to cache), but i am willing to make changes to allow this file to be used.

Signed-off-by: Jacob Hrbek <kreyren@member.fsf.org>