Only run xdg-* functions in deb postint, prerm if the requried dirs exist
Some distributions (debian, arch, maybe others) do not always create
the system-level /usr{,/local}/share/destop-directories paths which are
required for xdg-desktop-menu and similar programs to work correctly.

In addition, the prerm script uses the -e bash option (exit on non-zero
status).  However, the postinst script does not.

This can lead to the situation where the deb package can be installed,
but cannot be uninstalled. (upgrading still works)

This commit adds checks to the postinst and prerm scripts to only run the
xdg programs if at least one of the required directories exists.

This commit also adds the -e bash option to the postinst script, for
consistency.