[cpack] Refactor architecture tag logic
Changes the logic to append nothing to the package name when `dpkg`
is not present on the system.  This is a change from current
behavior which default assumes "i386".  The inline comment of this
section makes it look like the intent was never to use this default
value.  dpkg-based systems should not be affected, but rpm-based
systems will no longer have "i386" appear regardless of
architecture.