[cpack] Fix unsanitized package naming
This fixes bad logic in creating `CPACK_PACKAGE_FILE_NAME` that
would cause it to error out for Linux distributions with whitespace
in their `NAME` (as defined in `/etc/os-release`) which Fedora 35
and Debian do.  This updates that section to look for the `ID`
field instead which per `man 5 os-release` is guaranteed not to
have whitespace and be lowercase.  This should not change behavior
on Ubuntu & Debian but allows RPMs to build correctly on Fedora 35.