trashMemory() only calls the last std::memset().
Since the output of the first two calls to memset are not read before
the last call, those first two calls can be skipped by the compiler.
For example, clang-15 only calls the last memset().

Combine the existing approach with OpenSSL's use of calling memset()
through a volatile function pointer.  On Win32, use SecureZeroMemory()
for zeroisation, since it is designed for that purpose.

https://github.com/openssl/openssl/blob/master/crypto/mem_clr.c