Don't build the .pot/.po files in the source tree.
The "update-pot", "update-pos", and "update-translations" targets can be
used to update the in-tree files (the latter target includes the first
two).  This breaks the circular dependencies inherent when doing in-place
updates of the .po files in the source tree.

We add the USE_EXISTING_TRANSLATIONS option to use the existing .po files
when building the translation DLLs.  If false, the DLLs will be built from
updated .po files created in the build tree.  This option is enabled by
default.