[GH1041] - check signature of app-specific dlls
When the pwsafe.exe executable is signed, we now check that the pws_osk and pws_at dlls are signed as well.
If not, we just fail to load them.

Note that by restricting the check to when the exe is signed, we avoid the need to sign debug versions of the dlls, which is useful
for development.

If someone relaese an unsigned version of the exe, then all security bets are off for that, so no sense in checking there...