macOS: Yubikey serial number display problem
A late problem noticed in the Yubikey dialog on macOS:
With wxWidgets 3.2.2.1, the box for the Yubikey serial number does not resize,
it stays one character wide.  This does not happen with wx 3.2.4, however, that
version has a more serious regression.  As a workaround, adding an initial string
of zeros forces the box large enough for the 8 digit serial number.